% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationRange.R
\name{getCorrelationRange}
\alias{getCorrelationRange}
\title{Compute the correlation range values for all genes in the gene-gene correlation matrix.}
\usage{
getCorrelationRange(correlation_matrix)
}
\arguments{
\item{correlation_matrix}{gene-gene correlation matrix}
}
\value{
list of p-values, adjusted p-values and correlation ranges for each gene
}
\description{
Compute the correlation range values for all genes in the gene-gene correlation matrix.
}
