% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilteredData.R
\name{getFilteredData}
\alias{getFilteredData}
\title{Filter the dataset by removing lowly expressed genes and mitochondrial, spike-in and ribosomal genes}
\usage{
getFilteredData(data, min.cells = 0.05 * ncol(data))
}
\arguments{
\item{data}{gene expression matrix}

\item{min.cells}{gene expression matrix}
}
\value{
filtered gene-expression matrix
}
\description{
Filter the dataset by removing lowly expressed genes and mitochondrial, spike-in and ribosomal genes
}
