% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptimalFeatureSet.R
\name{getOptimalFeatureSet}
\alias{getOptimalFeatureSet}
\title{Determine the optimal feature set using Compactness Index (CI)}
\usage{
getOptimalFeatureSet(log.data, ordered.genes)
}
\arguments{
\item{log.data}{log-transformed gene-expression matrix}

\item{ordered.genes}{genes ordered after stepwise regression}
}
\value{
optimal feature set
}
\description{
Determine the optimal feature set using Compactness Index (CI)
}
