% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormalize.R
\name{logNormalize}
\alias{logNormalize}
\title{Log-transform and normalize data by sequencing depth}
\usage{
logNormalize(raw.data, scale.factor = 10000)
}
\arguments{
\item{raw.data}{raw gene expression matrix}

\item{scale.factor}{scaling factor for normalization}
}
\value{
log-normalized gene expression matrix
}
\description{
Log-transform and normalize data by sequencing depth
}
