% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_genes-data.R
\docType{data}
\name{pseudo_genes}
\alias{pseudo_genes}
\title{Pseudo-genes}
\format{An object of class \code{character} of length 14627.}
\source{
\href{http://www.ensembl.org}{ENSEMBL}
}
\usage{
data(pseudo_genes)
}
\description{
Pseudo-genes
}
\keyword{datasets}
