% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUBStepR.R
\name{DUBStepR}
\alias{DUBStepR}
\title{DUBStepR - Obtain a list of feature genes to characterise cell types}
\usage{
DUBStepR(
  input.data,
  min.cells = 100,
  k = 10,
  smooth = F,
  num.pcs = 15,
  error = 0
)
}
\arguments{
\item{input.data}{input gene expression matrix (genes x cells)}

\item{min.cells}{minimum number of cells to filter genes out and smooth data over}

\item{k}{number of nearest neighbours. Default is 10 - no smoothing.}

\item{smooth}{Boolean determining whether smoothing is necessary or not.}

\item{num.pcs}{number of principal components to represent sc data. Default is 15.}

\item{error}{Acceptable error margin for kNN computation. Default is 0.}
}
\value{
Returns optimal feature set
}
\description{
DUBStepR - Obtain a list of feature genes to characterise cell types
}
\author{
ranjanb
}
