% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUBStepR.R
\name{DUBStepR}
\alias{DUBStepR}
\title{DUBStepR - Obtain a list of feature genes to characterise cell types}
\usage{
DUBStepR(log.data, is_noisy = FALSE)
}
\arguments{
\item{log.data}{normalised log-transformed genes x cells single-cell RNA-seq data}

\item{is_noisy}{TRUE if data is noisy, FALSE otherwise.}
}
\value{
Returns a list of all feature genes and the optimal feature set
}
\description{
DUBStepR - Obtain a list of feature genes to characterise cell types
}
\author{
ranjanb
}
