% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUBStepR.R
\name{DUBStepR}
\alias{DUBStepR}
\title{DUBStepR - Obtain a list of feature genes to characterise cell types}
\usage{
DUBStepR(raw.data, log.normalize = T, min.cells = 0.05 *
  ncol(raw.data), k = 10, num.pcs = 15)
}
\arguments{
\item{raw.data}{raw gene expression matrix (genes x cells)}

\item{log.normalize}{boolean indicating whether input matrix needs to be log-normalized. Default is TRUE.}

\item{min.cells}{minimum number of cells to filter genes out and smooth data over}

\item{k}{number of nearest neighbours. Default is 10}

\item{num.pcs}{number of principal components to represent sc data. Default is 15.}
}
\value{
Returns optimal feature set
}
\description{
DUBStepR - Obtain a list of feature genes to characterise cell types
}
\author{
ranjanb
}
