% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGGC.R
\name{getGGC}
\alias{getGGC}
\title{Compute the correlation range values for all genes in the gene-gene correlation matrix}
\usage{
getGGC(log.data, is_noisy = FALSE)
}
\arguments{
\item{log.data}{log-transformed gene-expression matrix}

\item{correlation_matrix}{gene-gene correlation matrix}
}
\value{
list of genes with their z-transformed correlation range values
}
\description{
Compute the correlation range values for all genes in the gene-gene correlation matrix
}
