% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNNSmoothing.R
\name{kNNSmoothing}
\alias{kNNSmoothing}
\title{k-NN Smoothing of gene expression data}
\usage{
kNNSmoothing(log.filt.data)
}
\arguments{
\item{log.filt.data}{filtered and normalised log-transformed genes x cells single-cell RNA-seq data matrix}
}
\value{
k-NN Smoothed data matrix
}
\description{
k-NN Smoothing of gene expression data
}
