% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStepwiseReg.R
\name{runStepwiseReg}
\alias{runStepwiseReg}
\title{Run step-wise regression to order the features}
\usage{
runStepwiseReg(ggc)
}
\arguments{
\item{ggc}{gene-gene correlation matrix}
}
\value{
list of ordered genes
}
\description{
Run step-wise regression to order the features
}
